services/backend/pyproject.toml
[project]
name = "datarush-backend"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
 "argon2-cffi>=23.1.0",
 "celery>=5.4.0",
 "colorlog>=6.9.0",
 "django-cors-headers>=4.6.0",
 "django-environ>=0.11.2",
 "django-extensions>=3.2.3",
 "django-guid>=3.5.0",
 "django-health-check>=3.18.3",
 "django-minio-storage>=0.5.7",
 "django-ninja>=1.3.0",
 "django-pagedown>=2.2.1",
 "django-stubs-ext>=5.1.3",
 "gunicorn>=23.0.0",
 "httpx>=0.28.1",
 "pillow>=11.1.0",
 "psycopg2-binary>=2.9.10",
 "pydantic>=2.10.5",
 "pyjwt>=2.10.1",
 "python-gettext>=5.0",
 "python-json-logger>=3.2.1",
 "pytz>=2024.2",
 "redis>=5.2.1",
]

[dependency-groups]
dev = [
    "coverage>=7.6.12",
    "django-debug-toolbar>=4.4.6",
    "django-stubs[compatible-mypy]>=5.1.3",
    "mypy>=1.15.0",
    "ruff>=0.9.3",
]

[tool.ruff]
builtins = []
cache-dir = ".ruff_cache"
exclude = [
 ".bzr",
 ".direnv",
 ".eggs",
 ".git",
 ".git-rewrite",
 ".hg",
 ".mypy_cache",
 ".nox",
 ".pants.d",
 ".pytype",
 ".ruff_cache",
 ".svn",
 ".tox",
 ".venv",
 "__pypackages__",
 "_build",
 "buck-out",
 "dist",
 "migrations",
 "node_modules",
 "venv",
]
extend-exclude = []
extend-include = []
fix = false
fix-only = false
force-exclude = true
include = ["*.py", "*.pyi", "*.ipynb", "**/pyproject.toml"]
indent-width = 4
line-length = 79
namespace-packages = []
output-format = "full"
preview = false
required-version = ">=0.8.4"
respect-gitignore = true
show-fixes = true
src = [".", "src"]
target-version = "py310"
unsafe-fixes = false

[tool.ruff.analyze]
detect-string-imports = true
direction = "Dependencies"
exclude = []
include-dependencies = {}
preview = false

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 79
exclude = []
indent-style = "space"
line-ending = "lf"
preview = false
quote-style  = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
allowed-confusables = ["ℹ"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
exclude = ["tests.py"]
explicit-preview-rules = false
extend-fixable = []
extend-per-file-ignores = {}
extend-safe-fixes = []
extend-select = []
extend-unsafe-fixes = []
external = []
fixable = ["ALL"]
ignore = [
    "ARG",
    "D",
    "ANN401",
    "COM812",
    "DJ001",
    "DJ007",
    "FBT001",
    "FBT002",
    "N813",
    "PLR2004",
    "PT009",
    "PT027",
    "RUF001",
    "S311",
]
logger-objects = []
per-file-ignores = {}
preview = false
select = ["ALL"]
task-tags = ["TODO", "FIXME", "HACK", "WORKOUT"]
typing-modules = []
unfixable = []

[tool.ruff.lint.pylint]
max-args = 6

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
strict = false
show_error_context = false
no_implicit_optional = false

[tool.django-stubs]
django_settings_module = "config.settings"
strict_settings = false

[tool.coverage.run]
omit = [
    "manage.py",
    "config/wsgi.py",
    "config/asgi.py",
    "config/urls.py",
    "config/settings.py",
    "config/handlers.py",
    "config/errors.py",
]

[tool.coverage.report]
skip_covered = true
